Refactor useTable
