+++ src/apps/explorer/components/ActiveSolversTableWidget/index.tsx
@@ -4,12 +4,12 @@ import { ActiveSolversTableContext, BlockchainNetwork } from './context/ActiveSo
 import { useNetworkId } from 'state/network'
 import { useFlexSearch } from 'hooks/useFlexSearch'
 import { Solver, useGetSolvers } from 'hooks/useGetSolvers'
+import { TableState } from 'hooks/useTable'
 import { ActiveSolversTableWithData } from 'apps/explorer/components/ActiveSolversTableWidget/ActiveSolversTableWithData'
 import { ConnectionStatus } from 'components/ConnectionStatus'
 import { CardRow } from 'components/common/CardRow'
 import TablePagination, { PaginationWrapper } from '../common/TablePagination'
 import { DropdownDirection } from '../common/Dropdown'
-import { TableState } from './useTable'
 import { media } from 'theme/styles/media'
 import CowLoading from 'components/common/CowLoading'
 import { EmptyItemWrapper } from 'components/common/StyledUserDetailsTable'
