+++ src/apps/explorer/components/TokensTableWidget/index.tsx
@@ -4,13 +4,13 @@ import { BlockchainNetwork, TokensTableContext } from './context/TokensTableCont
 import { useNetworkId } from 'state/network'
 import { Token, useGetTokens } from 'hooks/useGetTokens'
 import { useFlexSearch } from 'hooks/useFlexSearch'
+import { useTable } from 'hooks/useTable'
 import { TokensTableWithData } from 'apps/explorer/components/TokensTableWidget/TokensTableWithData'
 import { TabItemInterface } from 'components/common/Tabs/Tabs'
 import ExplorerTabs from 'apps/explorer/components/common/ExplorerTabs/ExplorerTabs'
 import TablePagination from 'apps/explorer/components/common/TablePagination'
 import { ConnectionStatus } from 'components/ConnectionStatus'
 import { TabList } from 'components/common/Tabs/Tabs'
-import { useTable } from './useTable'
 import { TableSearch } from 'components/common/TableSearch/TableSearch'
 import { media } from 'theme/styles/media'
 import CowLoading from 'components/common/CowLoading'
