+++ src/apps/explorer/pages/Solver/index.tsx
@@ -3,9 +3,9 @@ import { useHistory } from 'react-router-dom'
 import { SupportedChainId } from '@cowprotocol/cow-sdk'
 import { useQuery } from 'hooks/useQuery'
 import { Solver as SolverType } from 'hooks/useGetSolvers'
+import { TableState, useTable } from 'hooks/useTable'
 import { useNetworkId } from 'state/network'
 import { ActiveSolversTableContext } from 'apps/explorer/components/ActiveSolversTableWidget/context/ActiveSolversTableContext'
-import { useTable, TableState } from 'apps/explorer/components/ActiveSolversTableWidget/useTable'
 import { TableSearch } from 'components/common/TableSearch/TableSearch'
 import TablePagination from 'apps/explorer/components/common/TablePagination'
 import ActiveSolversTableWidget from 'apps/explorer/components/ActiveSolversTableWidget'
