+++ test/hooks/useTable.test.tsx
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react-hooks'
-import { useTable } from 'apps/explorer/components/OrdersTableWidget/useTable'
+import { useTable } from 'hooks/useTable'
 
 describe('when rendered', () => {
   it('returns current initial value', () => {
